djangofiltertest/requirements/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
bcrypt==3.1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
decorator==4.3.0
Django==2.0
django-appconf==1.0.2
django-compressor==2.2
django-cors-headers==2.4.0
django-extensions==2.1.0
django-filter==2.0.0
django-model-utils==3.1.2
django-test-without-migrations==0.6
djangorestframework==3.8.2
djangorestframework-serializer-extensions==0.6.0
hashids==1.2.0
httpie==1.0.2
idna==2.8
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
Markdown==2.6.11
more-itertools==5.0.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.8.1
prompt-toolkit==2.0.7
ptyprocess==0.6.0
py==1.7.0
pycparser==2.18
Pygments==2.3.1
pytest==4.1.1
pytest-django==3.4.5
python-decouple==3.1
python-slugify==1.2.6
pytz==2018.5
rcssmin==1.0.6
requests==2.21.0
rjsmin==1.0.12
six==1.11.0
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.24.1
wcwidth==0.1.7
whitenoise==3.3.1
